+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to
 
 ### Changed
 
+- ♻️ (backend) Refactor room participant actions behind a facade #59
 - 🧹 (frontend) Resolved code smell #30
 - 🧹 (frontend) Refactor background processor init/state #33
 - 🧹 (frontend) Make `useMediaQuery` SSR-safe and update `matchMedia` #32